+++ multisig-wallet/ui/app/src/system/core/components/index.tsx
@@ -1,0 +1,3 @@
+export * from "./header";
+export * from "./layout";
+export * from "./please_connect";
