+++ multisig-wallet/ui/app/src/system/core/components/wallet_state.tsx
