+++ multisig-wallet/ui/app/src/system/core/contracts/MultisigContractAbi.d.ts
