+++ multisig-wallet/ui/app/src/system/core/contracts/common.d.ts
