+++ multisig-wallet/ui/app/src/system/core/contracts/factories/MultisigContractAbi__factory.ts
