+++ multisig-wallet/ui/app/src/system/core/hooks/useContract.ts
