+++ multisig-wallet/ui/app/src/system/core/hooks/useWallet.ts
