+++ multisig-wallet/ui/app/src/system/create/pages/create.tsx
@@ -1,0 +1,7 @@
+import { PleaseConnect } from "../../core/components";
+
+export function CreatePage() {
+    return (
+        <PleaseConnect />
+    );
+}
