+++ multisig-wallet/ui/app/src/system/create/pages/index.tsx
@@ -1,0 +1,1 @@
+export * from "./create";
